refactor(project-structure): reorganize project files
