+++ client/components/edit-commuter.js
@@ -7,9 +7,9 @@ import Form from 'react-formal'
 import {Map as LeafletMap, Marker, TileLayer} from 'react-leaflet'
 import yup from 'yup'
 
-import BackButton from '../containers/back-button'
-import FormalFieldGroup from './formal-fieldgroup'
-import Geocoder from './geocoder'
+import BackButton from '../containers/util/back-button'
+import FormalFieldGroup from './util/formal-fieldgroup'
+import Geocoder from './util/geocoder'
 import {actUponConfirmation, geocodeResultToState, geocodeYupSchema} from '../utils'
 import {pageview} from '../utils/analytics'
 import messages from '../utils/messages'
