+++ client/components/user-home.js
@@ -3,9 +3,9 @@ import {Button, ButtonGroup, Col, Grid, Row, Jumbotron} from 'react-bootstrap'
 import {BootstrapTable, TableHeaderColumn} from 'react-bootstrap-table'
 import {Link} from 'react-router'
 
-import ButtonLink from './button-link'
-import Icon from './icon'
-import HelpPopover from './help-popover'
+import ButtonLink from './util/button-link'
+import Icon from './util/icon'
+import HelpPopover from './util/help-popover'
 import {actUponConfirmation, arrayCountRenderer} from '../utils'
 import {pageview} from '../utils/analytics'
 import messages from '../utils/messages'
