+++ client/components/util/geocoder.js
@@ -3,7 +3,7 @@ import {ControlLabel, FormGroup} from 'react-bootstrap'
 import Form from 'react-formal'
 import SelectGeocoder from 'react-select-geocoder'
 
-import settings from '../utils/settings'
+import settings from '../../utils/settings'
 
 const boundary = {
   circle: {
