+++ client/containers/site-report.js
@@ -3,8 +3,8 @@ import {connect} from 'react-redux'
 import commuterActions from '../actions/commuter'
 import polygonActions from '../actions/polygon'
 import siteActions from '../actions/site'
-import makeDataDependentComponent from '../components/data-dependent-component'
-import SiteReport from '../components/site-report'
+import makeDataDependentComponent from '../components/util/data-dependent-component'
+import SiteReport from '../components/report'
 import {entityIdArrayToEntityArray} from '../utils/entities'
 
 function mapStateToProps (state, props) {
