+++ client/containers/util/back-button.js
@@ -1,7 +1,7 @@
 import {connect} from 'react-redux'
 import {goBack} from 'react-router-redux'
 
-import BackButton from '../components/back-button'
+import BackButton from '../../components/util/back-button'
 
 function mapStateToProps (state) {
   return {}
