+++ client/containers/util/login.js
@@ -1,7 +1,7 @@
 import {setAuth0User} from '@conveyal/woonerf/actions/user'
 import {connect} from 'react-redux'
 
-import Login from '../components/login'
+import Login from '../../components/util/login'
 
 function mapStateToProps (state, props) {
   return {}
